[vhost] Split server.cpp into endpoint/executor/server
server.cpp had grown to hold every class of the vhost server. Move them
into separate translation units, no behaviour changes.

* app_context.h - TAppContext, the server-wide state TServer derives from
  and shares with its endpoints.
* executor.h/.cpp - TExecutor and IRequestProcessor, the cookie contract
  it dispatches through. TExecutor no longer holds a TAppContext&: the
  only thing it needed from there was IServerStats, which is now a ctor
  argument.
* endpoint.h/.cpp - TEndpoint, TRequest and the per-request-type method
  traits, which stay private to endpoint.cpp.
* server.cpp - TServer only.
